EpsilonPrimeV2.1/requirements.txt
flask
python-dotenv
google-genai
chromadb
twilio
pydantic
requests
typing-extensions
psutil
langgraph
langchain-google-genai
langchain-core
smolagents
litellm
mcp
playwright
